Add jobService and refactor how system works
- add job service to manifest
- remove activity creation within geofence service
- clean-up unused code
- remove activity from geofencemanager
- created geofence jobservice to run location collectiion in background and upload data
