+++ libtelemetry/src/main/AndroidManifest.xml
@@ -9,6 +9,8 @@
         <service android:name="com.mapbox.android.telemetry.GeofenceIntentService"/>
         <service android:name=".LocationJobService"
                  android:permission="android.permission.BIND_JOB_SERVICE" />
+        <service android:name=".GeofenceJobService"
+                 android:permission="android.permission.BIND_JOB_SERVICE" />
     </application>
     <!--TODO Remove comment after analyzing the impact on the performance when adding SchedulerFlusherJobService-->
     <!--<application>-->
